feat(web): extract skill package validator helpers into skill-package-validator-lib
Move pure package parsing and submission-draft logic into a dedicated
-lib module with expanded Vitest coverage while preserving the public
@/lib/skill-package-validator import surface.

Closes #4286
